requirements.txt
apiron==5.1.0
bcrypt==4.3.0
beren==0.7.0
caldav==2.0.1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
configparser==7.2.0
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
fonttools==4.59.0
Genshi==0.7.9
icalendar==6.3.1
idna==3.10
kiwisolver==1.4.8
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
passlib==1.7.4
pendulum==3.1.0
pillow==11.3.0
polib==1.2.0
progressbar==2.2
psycopg2-binary==2.9.10
pycountry==20.7.3
pydicom==3.0.1
pyparsing==3.2.3
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-sql==1.6.0
python-stdnum==2.1
pytz==2025.2
PyWebDAV3-GNUHealth==0.13.0
qrcode==8.2
recurring-ical-events==3.8.0
relatorio==0.11.1
requests==2.32.4
simpleeval==1.0.3
six==1.17.0
tzdata==2025.2
unoconv==0.9.0
urllib3==2.5.0
vobject==0.9.9
Werkzeug==2.3.8
wrapt==1.17.2
x-wr-timezone==2.0.1
